Port divideWorld to Java

+++ src/main/java/net/avh4/rpg/generators/heroextant/Tile.java
@@ -2,7 +2,7 @@ package net.avh4.rpg.generators.heroextant;
 
 public class Tile {
 	TileType tileType = TileType.Undefined;
-	float elevation = 0;
+	double elevation = 0;
 	float wind = 0;
 	float rainfall = 0;
 	int waterSaturation = 0; // float
